[Chore] Migrated to new CarInput

+++ src/mirena_car.cpp
@@ -8,11 +8,11 @@ void MirenaCar::_bind_methods()
 	// GAS
 	ClassDB::bind_method(D_METHOD("get_gas"), &MirenaCar::get_gas);
 	ClassDB::bind_method(D_METHOD("set_gas", "_gas"), &MirenaCar::set_gas);
-	ADD_PROPERTY(PropertyInfo(Variant::INT, "gas"), "set_gas", "get_gas");
+	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "gas"), "set_gas", "get_gas");
 	// BRAKE
 	ClassDB::bind_method(D_METHOD("get_brake"), &MirenaCar::get_brake);
 	ClassDB::bind_method(D_METHOD("set_brake", "_brake"), &MirenaCar::set_brake);
-	ADD_PROPERTY(PropertyInfo(Variant::INT, "brake"), "set_brake", "get_brake");
+	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "brake"), "set_brake", "get_brake");
 	// STEER_ANGLE
 	ClassDB::bind_method(D_METHOD("get_steer_angle"), &MirenaCar::get_steer_angle);
 	ClassDB::bind_method(D_METHOD("set_steer_angle", "_steer_angle"), &MirenaCar::set_steer_angle);
@@ -44,19 +44,19 @@ void MirenaCar::_ros_ready()
 }
 
 // Getters and setters
-void MirenaCar::set_gas(uint8_t _gas)
+void MirenaCar::set_gas(float _gas)
 {
 	gas = _gas;
 }
-uint8_t MirenaCar::get_gas()
+float MirenaCar::get_gas()
 {
 	return gas;
 }
-void MirenaCar::set_brake(uint8_t _brake)
+void MirenaCar::set_brake(float _brake)
 {
 	brake = _brake;
 }
-uint8_t MirenaCar::get_brake()
+float MirenaCar::get_brake()
 {
 	return brake;
 }
